# Remove Result::$stack