--- a/src/main/php/lang/ast/emit/Result.class.php
+++ b/src/main/php/lang/ast/emit/Result.class.php
@@ -8,7 +8,6 @@ class Result implements Closeable {
   public $out;
   public $codegen;
   public $locals= [];
-  public $stack= [];
 
   /**
    * Starts a result stream.